Move controller to a separate file and move all logic to it
