+++ AirCasting/BluetoothSession/ReconnectionController.swift
@@ -49,50 +49,3 @@ class DefaultReconnectionObserver: ReconnectionController, BluetoothConnectionOb
         }
     }
 }
-
-enum UserTroggeredReconnectionError: Error {
-    case anotherActiveSessionInProgress
-    case failedToDiscoverCharacteristics
-    case failedToConnect
-    case deviceNotDiscovered
-    case airbeamConfigurationFailure
-}
-
-class UserTriggeredReconnectionController {
-    @Injected private var btScanner: BluetoothScanner
-    @Injected private var bluetootConnector: BluetoothConnectionHandler
-    @Injected private var activeSessionProvider: ActiveMobileSessionProvidingService
-    
-    func reconnectWithPeripheral(deviceUUID: String, completion: @escaping (Result<NewBluetoothManager.BluetoothDevice, UserTroggeredReconnectionError>) -> Void) {
-        guard activeSessionProvider.activeSession == nil else {
-            Log.info("Tried to resume session when there was another active session in progress")
-            completion(.failure(UserTroggeredReconnectionError.anotherActiveSessionInProgress))
-            return
-        }
-        var discoveredDevices: [NewBluetoothManager.BluetoothDevice] = []
-        btScanner.startScanning(scanningWindow: 10, onDeviceDiscovered: { discoveredDevices.append($0) }, onScanningFinished: {
-            guard let device = discoveredDevices.first(where: { $0.uuid == deviceUUID }) else {
-                completion(.failure(UserTroggeredReconnectionError.deviceNotDiscovered))
-                return
-            }
-            
-            self.bluetootConnector.connect(to: device, timeout: 10) { result in
-                switch result {
-                case .success:
-                    Log.info("Reconnected to a peripheral: \(device)")
-                    self.bluetootConnector.discoverCharacteristics(for: device, timeout: 10) { result in
-                        switch result {
-                        case .success:
-                            Log.info("Discovered characteristics for: \(device)")
-                            completion(.success((device)))
-                        case .failure(_):
-                            completion(.failure(UserTroggeredReconnectionError.failedToDiscoverCharacteristics))
-                        }
-                    }
-                case .failure(_):
-                    completion(.failure(UserTroggeredReconnectionError.failedToConnect))
-                }
-            }
-        })
-    }
-}
